steps/08_populate_imports_product_catalog_view_sp/requirements.txt
snowflake-snowpark-python
toml # for local development
pandas
